Add tests for managing rivals
